io: use 16 bit addresses
Although this is rarely used, we'll need this for the testsuite. At some
point we'll change to an address + mask model for device matching, but
for now this works.

+++ src/io.rs
@@ -2,13 +2,13 @@ use std::{collections::HashMap, fmt::Debug};
 
 pub trait Device {
     // XXX: maybe use upper 8 bits of address bus ? we don't need it for now
-    fn out(&self, addr: u8, val: u8) -> Result<(), String>;
-    fn input(&self, addr: u8) -> Result<u8, String>;
+    fn out(&self, addr: u16, val: u8) -> Result<(), String>;
+    fn input(&self, addr: u16) -> Result<u8, String>;
 }
 
 #[derive(Clone)]
 pub struct IO<'a> {
-    map: HashMap<u8, &'a dyn Device>,
+    map: HashMap<u16, &'a dyn Device>,
 }
 
 impl<'a> IO<'a> {
@@ -17,14 +17,14 @@ impl<'a> IO<'a> {
             map: HashMap::new(),
         }
     }
-    pub fn register(&mut self, addr: u8, dev: &'a mut dyn Device) {
+    pub fn register(&mut self, addr: u16, dev: &'a mut dyn Device) {
         self.map.insert(addr, dev);
     }
-    pub fn out(&self, addr: u8, val: u8) -> Result<(), String> {
+    pub fn out(&self, addr: u16, val: u8) -> Result<(), String> {
         let dev = self.map.get(&addr).ok_or("Out address not found")?;
         dev.out(addr, val)
     }
-    pub fn input(&self, addr: u8) -> Result<u8, String> {
+    pub fn input(&self, addr: u16) -> Result<u8, String> {
         let dev = self.map.get(&addr).ok_or("In address not found")?;
         dev.input(addr)
     }
